# lib/api-zod/src/generated/types/authResult.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * NearConnect API - local events discovery and community platform
 * OpenAPI spec version: 0.1.0
 */
import type { User } from './user';

export interface AuthResult {
  user: User;
  isNewUser: boolean;
}