# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * NearConnect API - local events discovery and community platform
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: number;
  name: string;
  /** @nullable */
  avatar?: string | null;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  lat?: number | null;
  /** @nullable */
  lng?: number | null;
  attributes?: string[];
  eventsHosted?: number;
  eventsJoined?: number;
  createdAt?: string;
}